BaseTest.kt: Update to setContent{ } due to update of PlutusWalletTheme(). SettingsValues.kt: Removed unneeded parameters. PreferenceHelper.kt: Removed Keys that were no longer in use. PWEnums.kt: Renamed from TransactionEnums.kt. Added comments for all enums. Updated names for *Action enums. SettingsUtils.kt: Removed preferences no longer used. Updated arguments passed to SettingsValues().
