+++ app/src/main/java/com/heyzeusv/plutuswallet/data/model/SettingsValues.kt
@@ -5,14 +5,11 @@ import java.text.DecimalFormat
 
 data class SettingsValues(
     var currencySymbol: String = "$",
-    var symbolSide: Boolean = true,
+    var currencySymbolSide: String = "left",
     var thousandsSymbol: Char = ',',
-    var decimalPlaces: Boolean = true,
     var decimalSymbol: Char = '.',
-    var dateFormat: Int = 0,
+    var decimalNumber: String = "yes",
     var decimalFormatter: DecimalFormat = DecimalFormat("#,##0.00"),
     var integerFormatter: DecimalFormat = DecimalFormat("#,###"),
-    var decimalNumber: String = "yes",
-    var currencySymbolSide: String = "left",
     var dateFormatter: DateFormat = DateFormat.getDateInstance(0)
 )
