+++ app/src/main/java/com/heyzeusv/plutuswallet/util/PreferenceHelper.kt
@@ -14,7 +14,8 @@ object PreferenceHelper {
     }
 
     /**
-     *  Puts a key value pair in shared prefs if doesn't exists, otherwise updates value on given [key].
+     *  Puts a key value pair in shared prefs if doesn't exists, otherwise updates value on
+     *  given [key].
      */
     operator fun SharedPreferences.set(key: Key, value: Any) {
 
@@ -51,19 +52,13 @@ object PreferenceHelper {
 
 // SharedPreferences Keys
 enum class Key(val key: String) {
-    KEY_CHART_CHANGED("key_chart_changed"),
     KEY_CURRENCY_SYMBOL_SIDE("key_currency_symbol_side"),
     KEY_CURRENCY_SYMBOL("key_currency_symbol"),
     KEY_DATE_FORMAT("key_date_format"),
-    KEY_DECIMAL_PLACES("key_decimal_places"),
     KEY_DECIMAL_NUMBER("key_decimal_number"),
     KEY_DECIMAL_SYMBOL("key_decimal_symbol"),
     KEY_LANGUAGE("key_language"),
-    KEY_LANGUAGE_CHANGED("key_language_changed"),
     KEY_MANUAL_LANGUAGE("key_manual_language"),
-    KEY_SYMBOL_SIDE("key_symbol_side"),
     KEY_THEME("key_theme"),
-    KEY_THEME_CHANGED("key_theme_changed"),
     KEY_THOUSANDS_SYMBOL("key_thousands_symbol"),
-    KEY_TRAN_LIST_CHANGED("key_tran_list_changed")
 }
