+++ app/src/test/java/com/heyzeusv/plutuswallet/ui/category/CategoryViewModelTest.kt
@@ -6,9 +6,9 @@ import com.heyzeusv.plutuswallet.data.FakeRepository
 import com.heyzeusv.plutuswallet.data.model.Category
 import com.heyzeusv.plutuswallet.data.model.DataDialog
 import com.heyzeusv.plutuswallet.ui.list.CategoryViewModel
-import com.heyzeusv.plutuswallet.util.DataListSelectedAction.CREATE
-import com.heyzeusv.plutuswallet.util.DataListSelectedAction.EDIT
-import com.heyzeusv.plutuswallet.util.DataListSelectedAction.DELETE
+import com.heyzeusv.plutuswallet.util.ListItemAction.CREATE
+import com.heyzeusv.plutuswallet.util.ListItemAction.EDIT
+import com.heyzeusv.plutuswallet.util.ListItemAction.DELETE
 import com.heyzeusv.plutuswallet.util.TransactionType.EXPENSE
 import com.heyzeusv.plutuswallet.util.TransactionType.INCOME
 import kotlinx.coroutines.ExperimentalCoroutinesApi
